--- Gmail.txt_example
@@ -1,5 +1,0 @@
-﻿john_doe@email.com:jdoe
-alice@email.net:alice
-bob@email.org:bob
-foo_bar@condemor.com:foobar
-chunkylover53@aol.com:chunky
