--- my_emails_list.txt_example
@@ -1,2 +1,0 @@
-aLiCe@email.net
-BoB@email.org
