+++ output.txt
@@ -1,0 +1,2 @@
+alice@email.net
+bob@email.org
